+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [[UNRELEASED](https://github.com/sysflow-telemetry/sf-apis/compare/0.3.0...HEAD)]
 
+### Changed
+
+- Update(ubi): Bumped UBI-minimal version to 8.4-210 in sysprint.
+
 ## [[0.3.0](https://github.com/sysflow-telemetry/sf-apis/compare/0.2.2...0.3.0)] - 2021-09-20
 
 ### Added